Move code to voicechat package
