+++ common-proxy/src/main/java/de/maxhenkel/voicechat/config/ProxyConfig.java
@@ -1,4 +1,4 @@
-package de.maxhenkel.voicechat.proxy.config;
+package de.maxhenkel.voicechat.config;
 
 import de.maxhenkel.configbuilder.ConfigBuilder;
 import de.maxhenkel.configbuilder.entry.ConfigEntry;
