+++ common-proxy/src/main/java/de/maxhenkel/voicechat/logging/VoiceChatLogger.java
@@ -1,4 +1,4 @@
-package de.maxhenkel.voicechat.proxy.logging;
+package de.maxhenkel.voicechat.logging;
 
 public interface VoiceChatLogger {
 
@@ -6,27 +6,27 @@ public interface VoiceChatLogger {
 
     boolean isEnabled(LogLevel level);
 
-    default void trace(String message, Object... args){
+    default void trace(String message, Object... args) {
         log(LogLevel.TRACE, message, args);
     }
 
-    default void debug(String message, Object... args){
+    default void debug(String message, Object... args) {
         log(LogLevel.DEBUG, message, args);
     }
 
-    default void info(String message, Object... args){
+    default void info(String message, Object... args) {
         log(LogLevel.INFO, message, args);
     }
 
-    default void warn(String message, Object... args){
+    default void warn(String message, Object... args) {
         log(LogLevel.WARN, message, args);
     }
 
-    default void error(String message, Object... args){
+    default void error(String message, Object... args) {
         log(LogLevel.ERROR, message, args);
     }
 
-    default void fatal(String message, Object... args){
+    default void fatal(String message, Object... args) {
         log(LogLevel.FATAL, message, args);
     }
